+++ resources/views/eventolist.blade.php
@@ -129,10 +129,13 @@
                             <a href="/evento/edit/{{$evento->id}}" style="text-decoration: none;">
                                 <span data-feather="edit"></span>
                             </a>
-                            &nbsp;
-                            <a href="/evento/delete/{{$evento->id}}" style="text-decoration: none;">
-                                <span data-feather="trash"></span>
-                            </a>
+                            @if(auth()->user()->grupo == 'admin')
+                                &nbsp;
+                                <a href="/evento/delete/{{$evento->id}}" style="text-decoration: none;">
+                                    <span data-feather="trash"></span>
+                                </a>
+                            @endif
+
                         </td>
                     </tr>
                 @endforeach
